Deleted unused classes/structs.
